pyproject.toml
[project]
name = "agentuity"
dynamic = ["version"]
description = "Agentuity SDK for Python"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE.md" }
dependencies = [
    "opentelemetry-api>=1.31.1",
    "opentelemetry-sdk>=1.31.1",
    "opentelemetry-instrumentation>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp>=1.31.1",
    "opentelemetry-exporter-otlp-proto-http>=1.31.1",
    "pyyaml>=6.0.2",
    "wrapt>=1.17.2",
    "aiohttp>=3.11.13",
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "aiohttp-sse>=2.2.0",
    "openlit>=1.33.19",
    "openai-agents>=0.0.3",
    "mail-parser>=4.1.2",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.11.1",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://agentuity.com"

[tool.setuptools.packages.find]
where = ["agentuity"]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"
full-commit = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
